refactor: change backend structure and add HTTP handlers, middleware, and tests for job analysis and rate limiting
