+++ apps/thread-dump-analyzer-tool/backend/internal/config/config.go
@@ -14,7 +14,7 @@
 // specific language governing permissions and limitations
 // under the License.
 
-package main
+package config
 
 import (
 	"fmt"
